Extract light color stuff into dedicated unit.
